Create README.md
This is a repository for Hax's Gamecube Controller mod code demonstrated on his YouTube and on Imgur:

https://www.youtube.com/watch?v=H4MXJt8UspQ
http://imgur.com/gallery/LFWDg
https://pastebin.com/YvB9BTp9

The code itself was rough to read and understand. It also hinted at more mods to come, but Hax has since abandoned hardware mods and moved to software mods through Gamecube Memory Card code injections. This project aims to clean up the original code and possibly implement more mods for the purposes of a tournament.

Notes: Using this REQUIRES NicoHoods "Nintendo" library built for the Arduino. I have included and modified this library slightly in the project.
https://github.com/NicoHood/Nintendo